ranks: use int40 (5-byte) instead of int32 for large-input viability
int32 ranks would overflow for the 100s-of-GB inputs the algorithm must
support; int40 covers +/-2^39 while still shrinking ranks below 8-byte long.
Applies to RunRecord, ranks_* files, and global_* merge output. HeapElement
keeps int64 fields (unpacked once per record) so the hot compare stays cheap.

+++ src/algorithm.h
@@ -2,13 +2,15 @@
 #define ALGORITHM_H
 
 #include <stdint.h>
+#include "utils.h"   // int40 + i40_load/i40_store
 
-// Rank values fit in int32 (total symbol count < 2^31 for supported inputs).
-// Storing ranks as int32 halves the on-disk runs_* footprint and the heap
-// element size that the merge stage streams through.
+// Rank values are stored as int40 (5 bytes) on disk and in streaming buffers.
+// This shrinks the runs_* footprint vs 8-byte long while staying valid for the
+// large (100s-of-GB) inputs the algorithm must support, where int32 would
+// overflow. count is a chunk-local run length (<= chunk_size <= 2^30).
 typedef struct run_triple {
-	int32_t currentRank;
-	int32_t nextRank;
+	int40 currentRank;
+	int40 nextRank;
 	int count;
 } RunRecord;
 
@@ -19,13 +21,13 @@ typedef struct rank_pos_pair {
 
 int generate_local_runs (char * rank_dir, char * runs_dir, int total_chunks, int chunk_id, int h,
                          long working_chunk_size,
-                         int32_t * current_ranks_buffer, int32_t * next_ranks_buffer,
+                         int40 * current_ranks_buffer, int40 * next_ranks_buffer,
                          int * sa_buffer, RunRecord * runs_buffer);
 
 int resolve_global_ranks (char *temp_dir );
 int update_local_ranks (char * rank_dir, char * temp_dir, int total_chunks, int chunk_id, int h,
                         long working_chunk_size,
-                        int32_t * buffer_current, int32_t * buffer_next,
-                        int * sa_buffer, long * updated_ranks);
+                        int40 * buffer_current, int40 * buffer_next,
+                        int * sa_buffer, int40 * updated_ranks);
 
 #endif
